1) Fixed adding links. Fixes #26 2) Changed update managing system. Fixes #25 and as it can gain data regularly it fixes #5

+++ User/States/SetUpdatesState.py
@@ -17,7 +17,7 @@ class SetUpdatesState(BasicState):
         if message == bot_strings.cancel_id:
             return BasicState.create_transition(StateTags.MAIN)
         try:
-            time = float(message)
+            time = int(message)
             if self.user.check_time_correct(time):
                 self.user.updates_duration = time
                 self.user.callback(bot_strings.set_updates_success)
